Refactor App into section components
